+++ demos/chinook/src/test/java/is/codion/demos/chinook/domain/ChinookTest.java
@@ -124,7 +124,7 @@ public class ChinookTest extends DomainTest {
 		@Override
 		public void modify(Entity entity) {
 			super.modify(entity);
-			if (entity.entityType().equals(Album.TYPE)) {
+			if (entity.type().equals(Album.TYPE)) {
 				entity.put(Album.TAGS, List.of("tag_one", "tag_two", "tag_three"));
 			}
 		}