+++ framework/domain/src/main/java/is/codion/framework/domain/entity/DefaultStringFactory.java
@@ -37,7 +37,7 @@ final class DefaultStringFactory implements Function<Entity, String>, Serializab
 
 	@Override
 	public String apply(Entity entity) {
-		return new StringBuilder(entity.entityType().name())
+		return new StringBuilder(entity.type().name())
 						.append(entity.definition().attributes().definitions().stream()
 										.map(AttributeDefinition::attribute)
 										.filter(entity::contains)