+++ framework/domain/src/test/java/is/codion/framework/domain/entity/EntitiesTest.java
@@ -218,7 +218,7 @@ public final class EntitiesTest {
 		Entity.Key key = entities.primaryKey(Master.TYPE, 10L);
 
 		Entity master = Entity.entity(key);
-		assertEquals(Master.TYPE, master.entityType());
+		assertEquals(Master.TYPE, master.type());
 		assertTrue(master.contains(Master.ID));
 		assertEquals(10L, master.get(Master.ID));