+++ framework/json-domain/src/main/java/is/codion/framework/json/domain/EntitySerializer.java
@@ -54,7 +54,7 @@ final class EntitySerializer extends StdSerializer<Entity> {
 	public void serialize(Entity entity, JsonGenerator generator, SerializerProvider provider) throws IOException {
 		requireNonNull(entity);
 		generator.writeStartObject();
-		generator.writeStringField("entityType", entity.entityType().name());
+		generator.writeStringField("entityType", entity.type().name());
 		generator.writeFieldName("values");
 		writeValues(entity, generator, entity.entrySet());
 		if (entity.modified()) {